task: split atmcfg file into smaller chunks
