+++ atmcfg/errors.go
@@ -1,0 +1,39 @@
+// Copyright 2022 MongoDB Inc
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package atmcfg // Package atmcfg import "go.mongodb.org/ops-manager/atmcfg"
+
+import (
+	"errors"
+	"fmt"
+)
+
+// newProcessNotFoundError returns an error if a process was not found.
+func newProcessNotFoundError(clusterName string, processesMap map[string]bool) error {
+	var processesNotFound []string
+	for k, v := range processesMap {
+		if !v {
+			processesNotFound = append(processesNotFound, k)
+		}
+	}
+
+	if len(processesNotFound) > 0 {
+		return fmt.Errorf("%w in %s: %v", ErrProcessNotFound, clusterName, processesNotFound)
+	}
+
+	return nil
+}
+
+// ErrProcessNotFound means the process was not found for the given cluster name.
+var ErrProcessNotFound = errors.New("process not found")
